--- utils.py
@@ -1,5 +1,0 @@
-import textwrap
-
-
-def reformat(text: str) -> str:
-    return textwrap.dedent(text).strip()
